requirements.txt
streamlit==1.30.0
pandas==2.1.4
numpy==1.26.3
Pillow==9.3.0
scikit-learn==1.3.2
openpyxl==3.1.2
seaborn==0.13.2
matplotlib==3.8.3